Fix indexing bug
We want to draw all the vertices i + r, where r is the square (mod p) of the residues
1, 2, ..., (p - 1) / 2 INCLUSIVE.  Previously, the upper bound had not been included,
causing some edges not to get drawn.

In light of this fix, we no longer need the 'visited' array to keep track of which
edges have already been drawn; all we have to do is ONLY draw the "forward" edges
(i.e., the edges ij where j > i).

Fixes #1