Add animation
This is done via setInterval().  Instead of having a double for loop in the
draw() method, draw() simply calls setInterval(), which in turn calls the
callback drawLine() at set intervals of time.

A "speed" parameter is now included in the Paley drawer object.  This is a
misnomer, since the higher the number the slower the graph will be drawn.

Fixes #5