Speed up paley.draw() by skipping edges that have been drawn
Previously, drawLine() was called exactly once for every pair of vertices.
If the pair had already been drawn, it would simply return.  Because of the
use of setInterval(), each call took a fixed amount of time.

Now, in each call to drawLine() we increment i and j UNTIL an edge has been drawn,
so we only end up calling drawLine() once for each edge in the Paley graph.

There are p(p - 1) / 4 edges in the Paley graph of order p, versus p(p - 1) / 2
pairs of vertices, so this decreases the time it takes to draw the graph by a
factor of 2.